Remove unnecessary SQL files from sbpapp directory
Numerous SQL files were removed from the sbpapp directory. They were redundant and no longer required as the project has been revamped. This helps in decluttering the repository, making the maintenance process smoother. It also makes it easier for other developers to understand the codebase and improve the maintainability of the project.